[gpgme] Backport from 1.6.0 the start of the agent on demand when signing S/MIME data. Contributes MER#1780
The signature code in gpgme, at line 1806 of engine-gpgsm.c, the line of the patch,
is using assuan to send through IPC the fingerprint of the signer. This assuan command
is running at line 665, and calling assuan_write_line() to send the signer fingerprint.

Then, it is looping through a assuan_read_line() function waiting for reply.
If no agent is currently running, before returning status OK, the read_line command
is returning a "S PROGRESS starting_agent" answer. This answer is caught at line
706 by the parse_status() function, which return a non zero value.

Passing NULL (i.e. no status_fnc) makes the code returns with an error because there
is no handler for non zero status case. The patch is thus providing a status handler
which acknowledges that "S PROGRESS starting_agent" is a waiting status not an error
one and a new call to read_line should be done, waiting for OK status.